+++ CHANGELOG.md
@@ -4,7 +4,11 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
-- Feature: Create `TeamCard` component.
+- Feature: Create `TeamCard` and sub-components:
+  - `TeamCard.Icons`
+  - `TeamCard.Image`
+  - `TeamCard.PhoneIcon`
+  - `TeamCard.Title`
 - Feature: Create `PhoneIcon` and `MailIcon` components.
 - Fix: Fix `ExpandBoard.Button` arrow styles.
 - Fix: Add `box-sizing` CSS rule to `RewardCard` component.
